json tests

+++ lib/features/number_trivia/domain/entities/number_trivia.dart
@@ -1,9 +1,8 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:equatable/equatable.dart';
 
 class NumberTriviaEntity extends Equatable {
   final String text;
-  final int number;
+  final double number;
   const NumberTriviaEntity({
     required this.text,
     required this.number,
